AWSMTestCase - Move to own file.
Some spring cleaning ...
